+++ src/cutest.c
@@ -1579,302 +1579,6 @@ error:
 }
 
 ///////////////////////////////////////////////////////////////////////////////
-// Option parser
-///////////////////////////////////////////////////////////////////////////////
-
-typedef enum cutest_optparse_argtype
-{
-    CUTEST_OPTPARSE_NONE,
-    CUTEST_OPTPARSE_REQUIRED,
-    CUTEST_OPTPARSE_OPTIONAL,
-} cutest_optparse_argtype_t;
-
-typedef struct cutest_optparse_long_opt
-{
-    const char* longname;
-    int                         shortname;
-    cutest_optparse_argtype_t   argtype;
-} cutest_optparse_long_opt_t;
-
-typedef struct cutest_optparse
-{
-    char** argv;
-    int                         permute;
-    int                         optind;
-    int                         optopt;
-    char* optarg;
-    char                        errmsg[64];
-    unsigned long               subopt;
-} cutest_optparse_t;
-
-#define OPTPARSE_MSG_INVALID "invalid option"
-#define OPTPARSE_MSG_MISSING "option requires an argument"
-#define OPTPARSE_MSG_TOOMANY "option takes no arguments"
-
-static int _cutest_optparse_is_dashdash(const char* arg)
-{
-    return arg != 0 && arg[0] == '-' && arg[1] == '-' && arg[2] == '\0';
-}
-
-static int _cutest_optparse_is_shortopt(const char* arg)
-{
-    return arg != 0 && arg[0] == '-' && arg[1] != '-' && arg[1] != '\0';
-}
-
-static int _cutest_optparse_is_longopt(const char* arg)
-{
-    return arg != 0 && arg[0] == '-' && arg[1] == '-' && arg[2] != '\0';
-}
-
-static int _cutest_optparse_longopts_end(const cutest_optparse_long_opt_t* longopts, int i)
-{
-    return !longopts[i].longname && !longopts[i].shortname;
-}
-
-static int _cutest_optparse_longopts_match(const char* longname, const char* option)
-{
-    const char* a = option, * n = longname;
-    if (longname == 0)
-        return 0;
-    for (; *a && *n && *a != '='; a++, n++)
-        if (*a != *n)
-            return 0;
-    return *n == '\0' && (*a == '\0' || *a == '=');
-}
-
-static char* _cutest_optparse_longopts_arg(char* option)
-{
-    for (; *option && *option != '='; option++);
-    if (*option == '=')
-    {
-        return option + 1;
-    }
-    return 0;
-}
-
-static void _cutest_optparse_from_long(const cutest_optparse_long_opt_t* longopts, char* optstring)
-{
-    char* p = optstring;
-    int i;
-    for (i = 0; !_cutest_optparse_longopts_end(longopts, i); i++)
-    {
-        if (longopts[i].shortname)
-        {
-            int a;
-            *p++ = (char)(longopts[i].shortname);
-            for (a = 0; a < (int)longopts[i].argtype; a++)
-            {
-                *p++ = ':';
-            }
-        }
-    }
-    *p = '\0';
-}
-
-static void _cutest_optparse_permute(cutest_optparse_t* options, int index)
-{
-    char* nonoption = options->argv[index];
-    int i;
-    for (i = index; i < options->optind - 1; i++)
-    {
-        options->argv[i] = options->argv[i + 1];
-    }
-    options->argv[options->optind - 1] = nonoption;
-}
-
-static int _cutest_optparse_argtype(const char* optstring, char c)
-{
-    int count = CUTEST_OPTPARSE_NONE;
-    if (c == ':')
-        return -1;
-    for (; *optstring && c != *optstring; optstring++);
-    if (!*optstring)
-        return -1;
-    if (optstring[1] == ':')
-        count += optstring[2] == ':' ? 2 : 1;
-    return count;
-}
-
-static int _cutest_optparse_error(cutest_optparse_t* options, const char* msg, const char* data)
-{
-    unsigned p = 0;
-    const char* sep = " -- '";
-    while (*msg)
-        options->errmsg[p++] = *msg++;
-    while (*sep)
-        options->errmsg[p++] = *sep++;
-    while (p < sizeof(options->errmsg) - 2 && *data)
-        options->errmsg[p++] = *data++;
-    options->errmsg[p++] = '\'';
-    options->errmsg[p++] = '\0';
-    return '?';
-}
-
-static int _cutest_optparse(cutest_optparse_t* options, const char* optstring)
-{
-    int type;
-    char* next;
-    char* option = options->argv[options->optind];
-    options->errmsg[0] = '\0';
-    options->optopt = 0;
-    options->optarg = 0;
-    if (option == 0) {
-        return -1;
-    }
-    else if (_cutest_optparse_is_dashdash(option)) {
-        options->optind++; /* consume "--" */
-        return -1;
-    }
-    else if (!_cutest_optparse_is_shortopt(option)) {
-        if (options->permute) {
-            int index = options->optind++;
-            int r = _cutest_optparse(options, optstring);
-            _cutest_optparse_permute(options, index);
-            options->optind--;
-            return r;
-        }
-        else {
-            return -1;
-        }
-    }
-    option += (options->subopt + 1);
-    options->optopt = option[0];
-    type = _cutest_optparse_argtype(optstring, option[0]);
-    next = options->argv[options->optind + 1];
-    switch (type) {
-    case -1: {
-        char str[2] = { 0, 0 };
-        str[0] = option[0];
-        options->optind++;
-        return _cutest_optparse_error(options, OPTPARSE_MSG_INVALID, str);
-    }
-    case CUTEST_OPTPARSE_NONE:
-        if (option[1]) {
-            options->subopt++;
-        }
-        else {
-            options->subopt = 0;
-            options->optind++;
-        }
-        return option[0];
-    case CUTEST_OPTPARSE_REQUIRED:
-        options->subopt = 0;
-        options->optind++;
-        if (option[1]) {
-            options->optarg = option + 1;
-        }
-        else if (next != 0) {
-            options->optarg = next;
-            options->optind++;
-        }
-        else {
-            char str[2] = { 0, 0 };
-            str[0] = option[0];
-            options->optarg = 0;
-            return _cutest_optparse_error(options, OPTPARSE_MSG_MISSING, str);
-        }
-        return option[0];
-    case CUTEST_OPTPARSE_OPTIONAL:
-        options->subopt = 0;
-        options->optind++;
-        if (option[1])
-            options->optarg = option + 1;
-        else
-            options->optarg = 0;
-        return option[0];
-    }
-    return 0;
-}
-
-static int _cutest_optparse_long_fallback(cutest_optparse_t* options, const cutest_optparse_long_opt_t* longopts, int* longindex)
-{
-    int result;
-    char optstring[96 * 3 + 1]; /* 96 ASCII printable characters */
-    _cutest_optparse_from_long(longopts, optstring);
-    result = _cutest_optparse(options, optstring);
-    if (longindex != 0) {
-        *longindex = -1;
-        if (result != -1) {
-            int i;
-            for (i = 0; !_cutest_optparse_longopts_end(longopts, i); i++)
-                if (longopts[i].shortname == options->optopt)
-                    *longindex = i;
-        }
-    }
-    return result;
-}
-
-static int cutest_optparse_long(cutest_optparse_t* options,
-    const cutest_optparse_long_opt_t* longopts, int* longindex)
-{
-    int i;
-    char* option = options->argv[options->optind];
-    if (option == 0) {
-        return -1;
-    }
-    else if (_cutest_optparse_is_dashdash(option)) {
-        options->optind++; /* consume "--" */
-        return -1;
-    }
-    else if (_cutest_optparse_is_shortopt(option)) {
-        return _cutest_optparse_long_fallback(options, longopts, longindex);
-    }
-    else if (!_cutest_optparse_is_longopt(option)) {
-        if (options->permute) {
-            int index = options->optind++;
-            int r = cutest_optparse_long(options, longopts, longindex);
-            _cutest_optparse_permute(options, index);
-            options->optind--;
-            return r;
-        }
-        else {
-            return -1;
-        }
-    }
-
-    /* Parse as long option. */
-    options->errmsg[0] = '\0';
-    options->optopt = 0;
-    options->optarg = 0;
-    option += 2; /* skip "--" */
-    options->optind++;
-    for (i = 0; !_cutest_optparse_longopts_end(longopts, i); i++) {
-        const char* name = longopts[i].longname;
-        if (_cutest_optparse_longopts_match(name, option)) {
-            char* arg;
-            if (longindex)
-                *longindex = i;
-            options->optopt = longopts[i].shortname;
-            arg = _cutest_optparse_longopts_arg(option);
-            if (longopts[i].argtype == CUTEST_OPTPARSE_NONE && arg != 0) {
-                return _cutest_optparse_error(options, OPTPARSE_MSG_TOOMANY, name);
-            } if (arg != 0) {
-                options->optarg = arg;
-            }
-            else if (longopts[i].argtype == CUTEST_OPTPARSE_REQUIRED) {
-                options->optarg = options->argv[options->optind];
-                if (options->optarg == 0)
-                    return _cutest_optparse_error(options, OPTPARSE_MSG_MISSING, name);
-                else
-                    options->optind++;
-            }
-            return options->optopt;
-        }
-    }
-    return _cutest_optparse_error(options, OPTPARSE_MSG_INVALID, option);
-}
-
-static void cutest_optparse_init(cutest_optparse_t * options, char** argv)
-{
-    options->argv = argv;
-    options->permute = 1;
-    options->optind = 1;
-    options->subopt = 0;
-    options->optarg = 0;
-    options->errmsg[0] = '\0';
-}
-
-///////////////////////////////////////////////////////////////////////////////
 // Timestamp
 ///////////////////////////////////////////////////////////////////////////////
 
@@ -2759,9 +2463,10 @@ static void _cutest_show_report(const cutest_porting_timespec_t* tv_total_start,
     _cutest_show_report_failed();
 }
 
-static void _cutest_setup_arg_pattern(char* user_pattern)
+static int _cutest_setup_arg_pattern(char* user_pattern)
 {
     g_test_ctx.filter.pattern = _cutest_str(user_pattern);
+    return 0;
 }
 
 /**
@@ -3218,7 +2923,25 @@ static void _cutest_prepare(void)
  */
 static int _cutest_setup(int argc, char* argv[], FILE* out, const cutest_hook_t* hook, int* b_exit)
 {
-    (void)argc;
+#define PARSER_LONGOPT_WITH_VALUE(OPT, FUNC)   \
+    do {\
+        int ret = -1;\
+        const char* opt = OPT;\
+        unsigned optlen = cutest_porting_strlen(opt);\
+        if (cutest_porting_strncmp(argv[i], opt, optlen) == 0) {\
+            if (argv[i][optlen] == '=') {\
+                ret = FUNC(argv[i] + optlen + 1);\
+            } else if (i < argc - 1) {\
+                ret = FUNC(argv[i + 1]);\
+                i++;\
+            }\
+            if (ret < 0) {\
+                _cutest_color_printf(CUTEST_COLOR_DEFAULT, "Invalid argument to `%s'\n", opt);\
+                return -1;\
+            }\
+            continue;\
+        }\
+    } while (0)
 
     cutest_porting_memset(&g_test_ctx, 0, sizeof(g_test_ctx));
     cutest_setup_once();
@@ -3227,88 +2950,53 @@ static int _cutest_setup(int argc, char* argv[], FILE* out, const cutest_hook_t*
     g_test_ctx.out = out;
     g_test_ctx.hook = hook;
 
-    enum test_opt
+    int i;
+    for (i = 0; i < argc; i++)
     {
-        test_list_tests = 1,
-        test_list_types,
-        test_filter,
-        test_also_run_disabled_tests,
-        test_repeat,
-        test_shuffle,
-        test_random_seed,
-        test_print_time,
-        test_break_on_failure,
-        help,
-    };
-
-    static cutest_optparse_long_opt_t longopts[] = {
-        { "test_list_tests",                test_list_tests,                CUTEST_OPTPARSE_NONE },
-        { "test_list_types",                test_list_types,                CUTEST_OPTPARSE_NONE },
-        { "test_filter",                    test_filter,                    CUTEST_OPTPARSE_REQUIRED },
-        { "test_also_run_disabled_tests",   test_also_run_disabled_tests,   CUTEST_OPTPARSE_NONE },
-        { "test_repeat",                    test_repeat,                    CUTEST_OPTPARSE_REQUIRED },
-        { "test_shuffle",                   test_shuffle,                   CUTEST_OPTPARSE_NONE },
-        { "test_random_seed",               test_random_seed,               CUTEST_OPTPARSE_REQUIRED },
-        { "test_print_time",                test_print_time,                CUTEST_OPTPARSE_REQUIRED },
-        { "test_break_on_failure",          test_break_on_failure,          CUTEST_OPTPARSE_NONE },
-        { "help",                           help,                           CUTEST_OPTPARSE_NONE },
-        { 0,                                0,                              0 },
-    };
-
-    cutest_optparse_t options;
-    cutest_optparse_init(&options, argv);
-
-    int option;
-    while ((option = cutest_optparse_long(&options, longopts, NULL)) != -1) {
-        switch (option) {
-        case test_list_tests:
-            _cutest_list_tests();
-            *b_exit = 1;
-            return 0;
-        case test_list_types:
-            _cutest_list_types();
-            *b_exit = 1;
-            return 0;
-        case test_filter:
-            _cutest_setup_arg_pattern(options.optarg);
-            break;
-        case test_also_run_disabled_tests:
-            g_test_ctx.mask.also_run_disabled_tests = 1;
-            break;
-        case test_repeat:
-            if (_cutest_setup_arg_repeat(options.optarg) != 0)
-            {
-                _cutest_color_printf(CUTEST_COLOR_DEFAULT, "Invalid argument to `--test_repeat'\n");
-                return -1;
-            }
-            break;
-        case test_shuffle:
-            g_test_ctx.mask.shuffle = 1;
-            break;
-        case test_random_seed:
-            if (_cutest_setup_arg_random_seed(options.optarg) != 0)
-            {
-                _cutest_color_printf(CUTEST_COLOR_DEFAULT, "Invalid argument to `--test_random_seed'\n");
-                return -1;
-            }
-            break;
-        case test_print_time:
-            if (_cutest_setup_arg_print_time(options.optarg) != 0)
-            {
-                _cutest_color_printf(CUTEST_COLOR_DEFAULT, "Invalid argument to `--test_print_time'\n");
-                return -1;
-            }
-            break;
-        case test_break_on_failure:
-            g_test_ctx.mask.break_on_failure = 1;
-            break;
-        case help:
+        if (cutest_porting_strcmp(argv[i], "--help") == 0
+            || cutest_porting_strcmp(argv[i], "-h") == 0)
+        {
             _cutest_print_encoded(s_test_help_encoded);
             *b_exit = 1;
             return 0;
-        default:
-            break;
         }
+
+        if (cutest_porting_strcmp(argv[i], "--test_list_tests") == 0)
+        {
+            _cutest_list_tests();
+            *b_exit = 1;
+            return 0;
+        }
+
+        if (cutest_porting_strcmp(argv[i], "--test_list_types") == 0)
+        {
+            _cutest_list_types();
+            *b_exit = 1;
+            return 0;
+        }
+
+        if (cutest_porting_strcmp(argv[i], "--test_also_run_disabled_tests") == 0)
+        {
+            g_test_ctx.mask.also_run_disabled_tests = 1;
+            continue;
+        }
+
+        if (cutest_porting_strcmp(argv[i], "--test_shuffle") == 0)
+        {
+            g_test_ctx.mask.shuffle = 1;
+            continue;
+        }
+
+        if (cutest_porting_strcmp(argv[i], "--test_break_on_failure") == 0)
+        {
+            g_test_ctx.mask.break_on_failure = 1;
+            continue;
+        }
+
+        PARSER_LONGOPT_WITH_VALUE("--test_filter", _cutest_setup_arg_pattern);
+        PARSER_LONGOPT_WITH_VALUE("--test_repeat", _cutest_setup_arg_repeat);
+        PARSER_LONGOPT_WITH_VALUE("--test_random_seed", _cutest_setup_arg_random_seed);
+        PARSER_LONGOPT_WITH_VALUE("--test_print_time", _cutest_setup_arg_print_time);
     }
 
     /* shuffle if necessary */
@@ -3318,6 +3006,8 @@ static int _cutest_setup(int argc, char* argv[], FILE* out, const cutest_hook_t*
     }
 
     return 0;
+
+#undef PARSER_LONGOPT_WITH_VALUE
 }
 
 static void _cutest_run_all_test_once(void)